+++ src/main/kotlin/de/reservationbear/eist/db/entity/Timeslot.kt
@@ -25,6 +25,6 @@ class Timeslot(
     @ColumnDefault("random_uuid()")
     @Type(type = "uuid-char")
     val id: UUID,
-    val timeslotFrom: Int,
-    val timeslotTo: Int
+    val timeslotFrom: Long,
+    val timeslotTo: Long
 )
